Merge branch 'feat/checkout-flow-optimization' into 'master'

Comprehensive rework of the ticketing lifecycle, checkout workflow, and documentation.

This merge integrates significant updates across the application, moving from legacy imperative flows to a modern, state-driven architecture for ticket purchasing and management.

Key Changes:

1.  Checkout & Payment Integration:
    - Refactored the checkout process into a 3-step React state machine for better UX.
    - Enhanced DOKU payment integration with support for pending result states.
    - Implemented a seamless public Event Details page with random featured event loading.

2.  Ticketing System Rework:
    - Standardized ticket and payment status nomenclature to English across the system.
    - Updated database schema for 'events' (FAQ, rules, contact info) and ticket enums.
    - Redesigned 'My Tickets' UI with a new color-coded status system and QR code security.
    - Added dedicated print-out views for purchased tickets.

3.  Testing & Infrastructure:
    - Modernized the backend test suite using object-oriented refactoring.
    - Centralized authentication helpers and database state management for tests.
    - Implemented robust integration tests for the financial and payment gateway logic.

4.  UI & Localization:
    - Fully localized the User and Organizer panels to English (Saved Events, Waiting Lists).
    - Refined navigation menus to remove redundant public links from administrative sidebars.
    - Fixed UI issues related to location text overflows and scan timestamp accuracy.

5.  Documentation:
    - Completed a total rewrite of README.md including the new technology stack, 
      RBAC credentials, and a detailed local development/payment testing guide.

+++ config/doku.php
@@ -32,6 +32,6 @@ return [
     /*
     | Expiry duration for Virtual Account payments (in minutes).
     */
-    'expiry_minutes' => env('DOKU_EXPIRY_MINUTES', 60),
+    'expiry_minutes' => (int) env('DOKU_EXPIRY_MINUTES', 60),
 
 ];
